+++ src/Visualizer.ts
@@ -7,11 +7,10 @@ import { VisualizerObject } from './templates/VisualizerObject'
 import { EventManagerInstance } from './system/EventManager'
 import { Effects } from './scene/Effects'
 import { UserDisplayGroup } from './scene/UserDisplayGroup'
-import { MainCircuit } from './scene/MainCircuit'
 import { VisualizerCamera } from './camera/VisualizerCamera'
 import { Recalculate } from './scene/Recalculates'
 import { getInitialData } from './utils/getInitialData'
-import { Circuit } from './scene/Circuit/CliantScript/Circuit'
+import { Circuit } from './scene/Circuit/Circuit'
 
 export class Visualizer {
   // singleton
